+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file. Be aware th
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.3.0] 2022-11-19
+### Added
+- New command to show or hide the pet bar: `/bbm pet`
+- New command to toggle mouseover: `/bbm toggle`
+
+### Fixed
+- The "Exit Vehicle" button now shows properly on mouseover
+
+### Removed
+- Removed old `/togglemo` command
+
 ## [1.2.0] 2022-11-16
 ### Added
 - The cooldown bling is now hidden